+++ ToyBox/Classes/Features/PartyTab/Stats/UnitModifyStatsFeature.cs
@@ -5,7 +5,6 @@ using Kingmaker.EntitySystem.Stats.Base;
 using Kingmaker.UnitLogic;
 using ToyBox.Infrastructure.Utilities;
 using UnityEngine;
-using Warhammer.SpaceCombat.StarshipLogic;
 
 namespace ToyBox.Features.PartyTab.Stats;
 
@@ -47,27 +46,31 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
             if (string.IsNullOrWhiteSpace(name)) {
                 name = stat.ToString();
             }
-            names.Add(name);
+            names.Add(name + " ");
         }
         return CalculateLargestLabelWidth(names, GUI.skin.label);
     });
-    private bool m_ShowDisclaimer = false;
+    private static string GetText(StatType stat) {
+        var name = LocalizedTexts.Instance.Stats.GetText(stat);
+        if (string.IsNullOrWhiteSpace(name)) {
+            name = stat.ToString();
+        }
+        return name;
+    }
+    private static readonly StatType[] m_Divider = [StatType.SkillAthletics, StatType.HitPoints];
     public void OnGui(BaseUnitEntity unit) {
         base.OnGui();
         if (IsEnabled) {
             using (HorizontalScope()) {
                 Space(25);
                 using (VerticalScope()) {
-                    UI.DisclosureToggle(ref m_ShowDisclaimer, m_TryToKeepThisFeatureActivatedAftLocalizedText.Orange());
-                    if (m_ShowDisclaimer) {
-                        using (HorizontalScope()) {
-                            Space(35);
-                            UI.Label("When this is turned off, the changed stats will still work in-game, but the respec UI might be very slightly buggy (e.g. +- values might be wrong). This is not a hard dependency as any issues can be fixed by respeccing the unit after disabling this feature/ToyBox.".Cyan(), Width(0.5f * EffectiveWindowWidth()));
-                        }
-                    }
+                    UI.Label("When this is turned off, the changed stats will be \"forgotten\" after reloading the save because base stats are recalculated on save load.".Orange(), Width(0.5f * EffectiveWindowWidth()));
                     foreach (StatType stat in Enum.GetValues(typeof(StatType))) {
                         if (Constants.WeirdStats.Contains(stat)) {
                             continue;
+                        }
+                        if (m_Divider.Contains(stat)) {
+                            Div.DrawDiv();
                         }
                         var modifiableValue = unit.Stats.GetStatOptional(stat);
                         var baseValue = 0;
@@ -76,15 +79,16 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
                             baseValue = modifiableValue.BaseValue;
                             modifiedValue = modifiableValue.ModifiedValue;
                         } else {
-                            // Note: We *could* support this by just not skipping the iteration here. 
+                            // Note: We *could* support this by just not skipping the iteration here.
+                            // And adding a nullable check to modifiableValue.m_Override != null
                             continue;
                         }
                         var change = 0;
                         using (HorizontalScope()) {
                             Space(10);
-                            var name = LocalizedTexts.Instance.Stats.GetText(stat);
-                            if (string.IsNullOrWhiteSpace(name)) {
-                                name = stat.ToString();
+                            var name = GetText(stat);
+                            if (modifiableValue.m_Override != null) {
+                                name += $" ({m_OverridenByLocalizedText}: {GetText(modifiableValue.m_Override.m_Type)})";
                             }
                             UI.Label(name, Width(m_LabelWidth));
                             _ = UI.Button("<", () => {
@@ -94,8 +98,6 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
                                     modifiedValue = modifiableValue.ModifiedValue;
                                 }
                                 change -= 1;
-                                modifiableValue.m_ActualBaseValue += change;
-                                modifiedValue += change;
                             });
                             UI.Label($" {modifiedValue} ".Bold().Orange(), Width(50 * Main.UIScale));
                             _ = UI.Button(">", () => {
@@ -105,8 +107,6 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
                                     modifiedValue = modifiableValue.ModifiedValue;
                                 }
                                 change += 1;
-                                modifiableValue.m_ActualBaseValue += change;
-                                modifiedValue += change;
                             });
                             Space(10);
                             var val = modifiedValue;
@@ -117,10 +117,9 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
                                     modifiedValue = modifiableValue.ModifiedValue;
                                 }
                                 change += pair.newContent - modifiableValue.ModifiedValue;
-                                modifiableValue.m_ActualBaseValue += change;
                             }, Width(75 * Main.UIScale));
                         }
-                        if (change > 0) {
+                        if (change != 0) {
                             if (InSaveSettings != null) {
                                 InSaveSettings.AppliedUnitStatChanges.TryGetValue(unit.UniqueId, out var dict);
                                 dict ??= [];
@@ -134,6 +133,8 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
                                 } else {
                                     dict[stat] = change;
                                 }
+                                modifiableValue.m_BaseValue = baseValue + change;
+                                modifiableValue.UpdateValue();
                                 InSaveSettings.AppliedUnitStatChanges[unit.UniqueId] = dict;
                                 InSaveSettings.Save();
                             }
@@ -143,13 +144,26 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
             }
         }
     }
+    [HarmonyPatch(typeof(BaseUnitEntity), nameof(BaseUnitEntity.GetStatBaseValue)), HarmonyPostfix]
+    private static void UnitHelper_CreatePreview_Patch(BaseUnitEntity __instance, StatType type, ref StatBaseValue __result) {
+        if (InSaveSettings?.AppliedUnitStatChanges.TryGetValue(__instance.UniqueId, out var changes) ?? false) {
+            if (changes.TryGetValue(type, out var change)) {
+                try {
+                    __result = new(__result.Value + change, __result.Enabled, __result.Forced);
+                } catch (Exception ex) {
+                    Error(ex);
+                }
+            }
+        }
+    }
     [HarmonyPatch(typeof(UnitHelper), nameof(UnitHelper.CreatePreview)), HarmonyPostfix]
     private static void UnitHelper_CreatePreview_Patch(BaseUnitEntity _this, ref BaseUnitEntity __result) {
         if (InSaveSettings?.AppliedUnitStatChanges.TryGetValue(_this.UniqueId, out var changes) ?? false) {
             foreach (var change in changes) {
                 try {
-                    var modifiableValue2 = __result.Stats.GetStatOptional(change.Key) ?? AddStat(change.Key, __result);
-                    modifiableValue2.m_ActualBaseValue += change.Value;
+                    var mV = __result.Stats.GetStatOptional(change.Key) ?? AddStat(change.Key, __result);
+                    mV.m_BaseValue += change.Value;
+                    mV.UpdateValue();
                 } catch (Exception ex) {
                     Error(ex);
                 }
@@ -170,4 +184,6 @@ public partial class UnitModifyStatsFeature : FeatureWithPatch, INeedContextFeat
 
     [LocalizedString("ToyBox_Features_PartyTab_Stats_UnitModifyStatsFeature_m_TryToKeepThisFeatureActivatedAftLocalizedText", "Try to keep this feature activated after using it (Click for Explanation)")]
     private static partial string m_TryToKeepThisFeatureActivatedAftLocalizedText { get; }
+    [LocalizedString("ToyBox_Features_PartyTab_Stats_UnitModifyStatsFeature_m_OverridenByLocalizedText", "Overriden by")]
+    private static partial string m_OverridenByLocalizedText { get; }
 }